refactor: InputController -> InputView 및 패키지 이동
