+++ src/main/java/racingcargame/view/InputView.java
@@ -1,4 +1,4 @@
-package racingcargame.controller;
+package racingcargame.view;
 
 import java.util.List;
 import java.util.Scanner;
@@ -6,15 +6,15 @@ import java.util.Scanner;
 import racingcargame.utils.NameInputValidator;
 import racingcargame.utils.RaceCountInputValidator;
 
-public class CarGameInputController {
-    private static final CarGameInputController INPUT_CONTROLLER = new CarGameInputController();
+public class InputView {
+    private static final InputView INPUT_CONTROLLER = new InputView();
     private static final Scanner SCANNER = new Scanner(System.in);
     private static final String CAR_NAME_DELIMITER = ",";
 
-    private CarGameInputController() {
+    private InputView() {
     }
 
-    public static CarGameInputController getInputController() {
+    public static InputView getInputController() {
         return INPUT_CONTROLLER;
     }
